requirements.txt
entsoe-py==0.6.8
slackweb==1.0.5
